:bug: Found iteration bug (deleted only half) of list
Deleting an element internally deletes the next node and modifies the reference so that the element is effectively removed, but the reference of the original "node to be deleted" now points to the next one. Therefore, iterating while (NULL != p_iter->next) and updating the reference p_iter = p_iter->next is a bad idea.